implement nmap file
